Refactor: Input과 관련된 컴포넌트 분리
